+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 - Default API & BOX url are api.signageos.io & box.signageos.io
 - Build production webpack will not start emulator
+- Live Reload webpack plugin of applet will trigger sos.onReady event
 
 ## [0.3.2] - 2019-09-05
 ### Fixed